feat: added dontAddToState prop field to InvokeActionForm

+++ src/components/visualEditor/ActionForms/invokeActionForm.tsx
@@ -18,7 +18,7 @@ import {InvokeActionProps, isState, ReactFlowContextProps} from "../../../types.
 export default function InvokeActionForm(props: {action: Action | undefined,
     setActions: Dispatch<SetStateAction<Action[]>>,
     onSubmit?: () => void,
-    noCategorySelect?: boolean}) {
+    noCategorySelect?: boolean, dontAddToState? :boolean }) {
 
     const context = useContext(ReactFlowContext) as ReactFlowContextProps;
     const {selectedNode,
@@ -182,9 +182,10 @@ export default function InvokeActionForm(props: {action: Action | undefined,
 
             // Check if category has changed and update the action in the state node if needed
             if (oldCategory !== selectedActionCategory as ActionCategory) {
-                //TODO provide optional dont add action to state prop.
-                stateOrStateMachineService.removeActionFromState(props.action, selectedNode.data);
-                stateOrStateMachineService.addActionToState(selectedNode.data, props.action, selectedActionCategory as ActionCategory);
+                if(!props.dontAddToState){
+                    stateOrStateMachineService.removeActionFromState(props.action, selectedNode.data);
+                    stateOrStateMachineService.addActionToState(selectedNode.data, props.action, selectedActionCategory as ActionCategory);
+                }
             }
 
             updatedAction = props.action;
@@ -194,7 +195,9 @@ export default function InvokeActionForm(props: {action: Action | undefined,
             updatedAction = new Action("newAction", ActionType.INVOKE);
             updatedAction.properties = invokeActionsProperties;
             //TODO provide optional dont add action to state prop.
-            stateOrStateMachineService.addActionToState(selectedNode.data, updatedAction, selectedActionCategory as ActionCategory);
+            if(!props.dontAddToState){
+                stateOrStateMachineService.addActionToState(selectedNode.data, updatedAction, selectedActionCategory as ActionCategory);
+            }
             onActionSubmit(updatedAction);
         }
